Removing DirectXTK Project Adding RTTR for reflection

+++ Albot/Code/Graphics/DebugRendering.cpp
@@ -206,13 +206,13 @@ void DebugRendering::render_debug_line_scene()
 	BufferUpdateDesc line_uniform_buffer_update_desc = {};
 	line_uniform_buffer_update_desc.m_buffer = m_line_uniform_structured_buffer;
 	line_uniform_buffer_update_desc.m_pSource = &m_line_vertex_uniform_data[0];
-	line_uniform_buffer_update_desc.m_size = sizeof(debug_line_count) * sizeof(LineVertexUniformData);
+	line_uniform_buffer_update_desc.m_size = static_cast<uint32_t>(sizeof(debug_line_count) * sizeof(LineVertexUniformData));
 	m_dxrenderer->cmd_update_buffer(line_uniform_buffer_update_desc);
 
 	BufferUpdateDesc line_vertex_buffer_update_desc = {};
 	line_vertex_buffer_update_desc.m_buffer = m_debug_line_vertex_buffer;
 	line_vertex_buffer_update_desc.m_pSource = &m_raw_line_vertices[0];
-	line_vertex_buffer_update_desc.m_size = sizeof(Vector3) * m_raw_line_vertices.size();
+	line_vertex_buffer_update_desc.m_size = static_cast<uint32_t>(sizeof(Vector3) * m_raw_line_vertices.size());
 	m_dxrenderer->cmd_update_buffer(line_vertex_buffer_update_desc);
